+++ package.json
@@ -2,14 +2,18 @@
   "name": "taskwave",
   "version": "1.0.0",
   "description": "A fast and efficient todo list application built with Bun.",
-  "main": "server.js",
+  "main": "src/backend/server.js",
   "scripts": {
-    "start": "bun run server.js",
-    "dev": "bun run --watch server.js"
+    "start": "bun src/backend/server.js",
+    "dev": "concurrently \"bun --watch src/backend/server.js\" \"vite --config vite.config.js\"",
+    "dev:frontend": "vite --config vite.config.js",
+    "build": "vite build --config vite.config.js",
+    "serve": "bun src/backend/server.js"
   },
   "author": "William Huang",
   "license": "MIT",
   "dependencies": {
+    "@vitejs/plugin-react": "^4.2.1",
     "bcryptjs": "^2.4.3",
     "chalk": "4.1.2",
     "cookie-parser": "^1.4.6",
@@ -18,6 +22,10 @@
     "jsonwebtoken": "^9.0.2",
     "mime": "^2.4.4",
     "mongoose": "^8.3.0",
-    "uuid": "^9.0.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "uuid": "^9.0.1",
+    "vite": "^5.2.8",
+    "concurrently": "^7.0.0" 
   }
 }
