+++ src/backend/config/auth.js
