+++ src/frontend/App.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const App = () => {
+  console.log("App rendering"); // Check if this logs in the console
+  return <h1>Hello, React!</h1>;
+};
+
+export default App;
