+++ src/frontend/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>React App</title>
+</head>
+
+<body>
+    <div id="root"></div>
+    <script type="module" src="/main.jsx"></script>
+</body>
+
+</html>
